+++ toolkit/cli/smart-contracts-commands/src/d_parameter.rs
@@ -1,5 +1,5 @@
 use crate::PaymentFilePath;
-use jsonrpsee::http_client::HttpClient;
+use ogmios_client::jsonrpsee::client_for_url;
 use partner_chains_cardano_offchain::await_tx::FixedDelayRetries;
 use partner_chains_cardano_offchain::d_param::upsert_d_param;
 use sidechain_domain::DParameter;
@@ -26,7 +26,7 @@ impl UpsertDParameterCmd {
 			num_permissioned_candidates: self.permissioned_candidates_count,
 			num_registered_candidates: self.registered_candidates_count,
 		};
-		let client = HttpClient::builder().build(self.common_arguments.ogmios_url)?;
+		let client = client_for_url(&self.common_arguments.ogmios_url).await?;
 
 		upsert_d_param(
 			self.genesis_utxo,