+++ changelog.md
@@ -8,6 +8,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 * `setup-main-chain-state` command now uses native Rust to upsert the D-Parameter and upsert permissioned candidates
 * Changed the `smart-contracts init-governance` command to `smart-contracts governance init`
+* smart-contracts commands and offchain tests now use WebSockets implementation of Ogmios client
 
 ## Removed
 
@@ -23,14 +24,16 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 * Command `smart-contracts reserve init`
 * Command `smart-contracts reserve create`
+* Command `smart-contracts reserve deposit`
 * Command `smart-contracts governance update`
+* Ogmios client backed by jsonrpsee `WsClient`
 
 # v1.4.0
 
 ## Changed
 
 * `genesis_utxo` and `registration_utxo` no longer have to have no native tokens.
-* Update ogmios to v6.9.0
+* Update dependencies containers to cardano-node 10.1.4, db-sync 13.6.0.4, ogmios 6.11.0 and kupo 2.10.0
 * Organized Rust sources into two directories: toolkit and node.
 * Implemented transaction balancing with CSL in offchain code.
 * Update offchain code dependencies: pallas, ulpc and cardano-serialization-lib.